Fix GCC 14 stringop-overflow warning in date library
- Updated buffer size calculation in read function to use unsigned long long
- Changed cast to unsigned long long for safer bounds checking
- Added changelog entry documenting the fix
- Resolves warning on Raspberry Pi Compute Module 5 with GCC 14

+++ external/date/include/date/date.h
@@ -6502,8 +6502,8 @@ read(std::basic_istream<CharT, Traits>& is, int a0, Args&& ...args)
 {
     if (a0 != -1)
     {
-        auto u = static_cast<unsigned>(a0);
-        CharT buf[std::numeric_limits<unsigned>::digits10+2u] = {};
+        auto u = static_cast<unsigned long long>(a0);
+        CharT buf[std::numeric_limits<unsigned long long>::digits10+2u] = {};
         auto e = buf;
         do
         {
